Fixing time

+++ userthread.c
@@ -378,7 +378,7 @@ void stub(void (*func)(void *), void *arg) {
     struct timeval time;
     gettimeofday(&time, NULL);
     //TODO: fix timing here
-    int time_in_mill = (time.tv_sec) * 1000 + (time.tv_usec) / 1000 ;
+    long time_in_mill = (time.tv_sec) * 1000 + (time.tv_usec) / 1000 ;
     return time_in_mill;
 }
